+++ requirements.jl
@@ -1,4 +1,5 @@
 using Pkg
 Pkg.add("Plots")
 Pkg.add("DifferentialEquations")
+Pkg.add("EzXML")
 Pkg.add(url="https://github.com/JuliaReliab/LookupTable.jl.git")